refactoring

+++ mysite/CharPage/views.py
@@ -10,11 +10,9 @@ from django.conf import settings
 from django.contrib import messages
 from django.core.exceptions import PermissionDenied
 from django.core.handlers.asgi import ASGIRequest
-from django.db import connection
 from django.db.models import QuerySet
 from django.http import Http404, HttpResponse, JsonResponse
 from django.shortcuts import redirect
-from django.test.utils import CaptureQueriesContext
 from django.utils import timezone
 from django.utils.decorators import method_decorator
 from django.views.decorators.cache import cache_page
@@ -128,10 +126,9 @@ class UniqueCharPageView(TitleMixin, TemplateView):
         dressing_rooms = CharModel.objects.all().only("creator", "last_update_time")
 
         for room in dressing_rooms:
-            if current_datetime - self.time_delta >= room.last_update_time:
+            if current_datetime - self.time_delta >= room.last_update_time and not room.creator:
                 # Если пользователь не прикреплён к модели удаляем модель
-                if not room.creator:
-                    room.delete()
+                room.delete()
 
     def _get_my_saved_rooms(self, creator: dict[str, str | User]) -> tuple[list[dict[str, str | int | bool]], QuerySet]:
         """Создание даты для отображения списка созданных комнат у текущего пользователя."""
@@ -207,32 +204,32 @@ class UniqueCharPageView(TitleMixin, TemplateView):
             else:
                 self.is_room_creator = self.creator_id == self.request.user
 
-            context["creating"] = getattr(self.dressing_room[0], "creating")
-            context["class"] = getattr(self.dressing_room[0], "char_class").lower()
-            context["race"] = RACES[getattr(self.dressing_room[0], "race")][0]
+            context["creating"] = self.dressing_room[0].creating
+            context["class"] = self.dressing_room[0].char_class.lower()
+            context["race"] = RACES[self.dressing_room[0].race][0]
             context["is_creator"] = self.is_room_creator
             context["race_image"] = (
-                f"../static/img/rass/{GENDERS[getattr(self.dressing_room[0], 'gender')]}/{RACES[getattr(self.dressing_room[0], 'race')][1]}.jpg"
+                f"../static/img/rass/{GENDERS[self.dressing_room[0].gender]}/{RACES[self.dressing_room[0].race][1]}.jpg"
             )
-            context["name"] = getattr(self.dressing_room[0], "char_name")
-            if eval(getattr(self.dressing_room[0], "proffesions"))[0] == 0:
-                if eval(getattr(self.dressing_room[0], "proffesions"))[1] == 0:
+            context["name"] = self.dressing_room[0].char_name
+            if eval(self.dressing_room[0].proffesions)[0] == 0:
+                if eval(self.dressing_room[0].proffesions)[1] == 0:
                     pass
                 else:
                     context["proffesion1_icon"] = (
-                        f"../static/img/Professions/large/{PROFFESIONS[eval(getattr(self.dressing_room[0], 'proffesions'))[1]][1]}.jpg"
+                        f"../static/img/Professions/large/{PROFFESIONS[eval(self.dressing_room[0].proffesions)[1]][1]}.jpg"
                     )
-                    context["proffesion1"] = PROFFESIONS[eval(getattr(self.dressing_room[0], "proffesions"))[1]][0]
+                    context["proffesion1"] = PROFFESIONS[eval(self.dressing_room[0].proffesions)[1]][0]
             else:
                 context["proffesion1_icon"] = (
-                    f"../static/img/Professions/large/{PROFFESIONS[eval(getattr(self.dressing_room[0], 'proffesions'))[0]][1]}.jpg"
+                    f"../static/img/Professions/large/{PROFFESIONS[eval(self.dressing_room[0].proffesions)[0]][1]}.jpg"
                 )
-                context["proffesion1"] = PROFFESIONS[eval(getattr(self.dressing_room[0], "proffesions"))[0]][0]
-                if not eval(getattr(self.dressing_room[0], "proffesions"))[1] == 0:
+                context["proffesion1"] = PROFFESIONS[eval(self.dressing_room[0].proffesions)[0]][0]
+                if eval(self.dressing_room[0].proffesions)[1] != 0:
                     context["proffesion2_icon"] = (
-                        f"../static/img/Professions/large/{PROFFESIONS[eval(getattr(self.dressing_room[0], 'proffesions'))[1]][1]}.jpg"
+                        f"../static/img/Professions/large/{PROFFESIONS[eval(self.dressing_room[0].proffesions)[1]][1]}.jpg"
                     )
-                    context["proffesion2"] = PROFFESIONS[eval(getattr(self.dressing_room[0], "proffesions"))[1]][0]
+                    context["proffesion2"] = PROFFESIONS[eval(self.dressing_room[0].proffesions)[1]][0]
 
             context["talents"] = self.dressing_room[0].talents.all()
             if not self.request.user.is_anonymous:
@@ -271,7 +268,7 @@ class UniqueCharPageView(TitleMixin, TemplateView):
             character_data = self.create_character_data()
 
             current_room = CharModel.objects.filter(room_id=self.room_id)
-            context["creating"] = getattr(current_room[0], "creating")
+            context["creating"] = current_room[0].creating
 
         context.update({"character_data": json.dumps(character_data)})
 
@@ -280,7 +277,7 @@ class UniqueCharPageView(TitleMixin, TemplateView):
     def render_to_response(self, context, **response_kwargs):
         if not context["creating"]:
             return redirect("createchar", self.room_id)
-        return super(UniqueCharPageView, self).render_to_response(context, **response_kwargs)
+        return super().render_to_response(context, **response_kwargs)
 
     def post(self, request: ASGIRequest, *args, **kwargs) -> JsonResponse:
         """Получает character_data с фронта и на её основе обновляет запись в БД."""
@@ -301,17 +298,15 @@ class UniqueCharPageView(TitleMixin, TemplateView):
 
         data: dict = json.loads(request.body)
         if self.is_room_creator:
-            if data.get("talent_id"):
-                if self.dressing_room[0].talents.count() < 2:
-                    talent = TalentsModel.objects.get(id=int(data["talent_id"]))
-                    talent.charmodel_set.add(self.dressing_room[0])
-                    return JsonResponse({"status": "data was successfully update talent"})
+            if data.get("talent_id") and self.dressing_room[0].talents.count() < 2:
+                talent = TalentsModel.objects.get(id=int(data["talent_id"]))
+                talent.charmodel_set.add(self.dressing_room[0])
+                return JsonResponse({"status": "data was successfully update talent"})
 
-            if data.get("deltalent_id"):
-                if self.dressing_room[0].talents.count() > 0:
-                    talent = TalentsModel.objects.get(id=int(data["deltalent_id"]))
-                    talent.charmodel_set.remove(self.dressing_room[0])
-                    return JsonResponse({"status": "data was successfully delete talent"})
+            if data.get("deltalent_id") and self.dressing_room[0].talents.count() > 0:
+                talent = TalentsModel.objects.get(id=int(data["deltalent_id"]))
+                talent.charmodel_set.remove(self.dressing_room[0])
+                return JsonResponse({"status": "data was successfully delete talent"})
 
             self.dressing_room.update(**data)
 
@@ -336,7 +331,7 @@ class CreateCharView(TitleMixin, TemplateView):
 
         self.room_id = self.kwargs["room_id"]
         current_room = CharModel.objects.get(room_id=self.kwargs["room_id"])
-        context["creating"] = getattr(current_room, "creating")
+        context["creating"] = current_room.creating
 
         return context
 
@@ -348,22 +343,22 @@ class CreateCharView(TitleMixin, TemplateView):
         if context["creating"]:
             return redirect("char_page_room", self.room_id)
         # Иначе если создатель существует и не равен текущему пользователю - запрещаем доступ к странице
-        if dressing_room[0].creator != self.request.user and dressing_room[0].creator != None:
-            raise PermissionDenied()
+        if dressing_room[0].creator != self.request.user and dressing_room[0].creator is not None:
+            raise PermissionDenied
 
-        return super(CreateCharView, self).render_to_response(context, **response_kwargs)
+        return super().render_to_response(context, **response_kwargs)
 
     def post(self, request: ASGIRequest, *args, **kwargs):
         data: dict = json.loads(request.body)
 
         current_room = CharModel.objects.get(room_id=self.kwargs["room_id"])
-        setattr(current_room, "char_name", data["Name"])
-        setattr(current_room, "proffesions", str(data["Proffesions"]))
-        setattr(current_room, "gender", data["Gender"])
-        setattr(current_room, "race", data["Race"])
-        setattr(current_room, "char_class", data["Class"])
-        setattr(current_room, "face", ",".join(map(str, data["Face_options"])))
-        setattr(current_room, "creating", True)
+        current_room.char_name = data["Name"]
+        current_room.proffesions = str(data["Proffesions"])
+        current_room.gender = data["Gender"]
+        current_room.race = data["Race"]
+        current_room.char_class = data["Class"]
+        current_room.face = ",".join(map(str, data["Face_options"]))
+        current_room.creating = True
         current_room.save()
 
         return JsonResponse({"status": "data was successfully saved"})
@@ -376,7 +371,7 @@ class CharListPageView(TitleMixin, TemplateView):
     title = "Мои персонажи"
 
     def get_context_data(self, **kwargs):
-        context = super(CharListPageView, self).get_context_data(**kwargs)
+        context = super().get_context_data(**kwargs)
         if not self.request.user.is_anonymous:
             context["all_chars"] = list(reversed(CharModel.objects.filter(creator=self.request.user, creating=True)))
 
@@ -386,7 +381,7 @@ class CharListPageView(TitleMixin, TemplateView):
         if self.request.user.is_anonymous:
             return redirect("char")
 
-        return super(CharListPageView, self).render_to_response(context, **response_kwargs)
+        return super().render_to_response(context, **response_kwargs)
 
     def post(self, request: ASGIRequest, *args, **kwargs):
         data: dict = json.loads(request.body)
@@ -399,5 +394,7 @@ class CharListPageView(TitleMixin, TemplateView):
                 obj_char.delete()
                 messages.success(self.request, "Персонаж успешно удален.")
                 return JsonResponse({"status": "The character has been successfully deleted"})
-            else:
-                return JsonResponse({"status": "access error"})
+
+            return JsonResponse({"status": "access error"})
+
+        return None
